backend_scheduler/requirements.txt
pandas==2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.3
sqlalchemy